requirements.txt
asgiref==3.8.1
Booktype==1.5
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==76.1.0
simplejson==3.20.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
